Update replay_lights.py
Added a parameter of switches to be excluded from the replay.  I just added a switch on my garage and a camera that gets enabled when the alarm is enabled.  I didn't want these include in the replay.  I also added some shelly switches to my home that ended up generating events without a new state and caused an exception, so I added code to just ignore these records in the database.