Update replay_lights.py
Code was pulling entity information from the events table.  The HA team stopped putting the information in this table because it was also stored in the states table.  I changed the code to pull the information from the states table.  I made the updates for both sqlite3 and MariaDB, however I don't have a MariaDB install so I wasn't able to test that part.